refactor: Update Transfer Points feature to match SQL specification
- Change Points from int to float64 (decimal) for precise monetary values
- Rename Points field to Amount in Transfer entities and models
- Add TransferredAt timestamp field to track exact transfer time
- Update database schema with proper column definitions:
  - transfers.amount: decimal(10,2) for precise currency handling
  - users.points: decimal(10,2) with default 0.00
  - Add proper foreign key constraints and indexes
- Update all GORM models with explicit column mappings and types
- Enhance Transfer repository with transaction safety
- Update API responses to include transferredAt field

Tested and verified:
✅ Transfer decimal amounts (150.25 points) with precision
✅ Proper deduction/addition of decimal points
✅ Transfer history with transferredAt timestamps
✅ Database schema matches SQL spec requirements
✅ All API endpoints working with new decimal format

+++ domain/repositories/user_repository.go
@@ -21,6 +21,6 @@ type UserRepository interface {
 	Update(id uint, user *entities.User) error
 	Delete(id uint) error
 	GetByEmail(email string) (*entities.User, error)
-	UpdatePoints(userID uint, points int) error
-	UpdatePointsWithTransaction(tx interface{}, userID uint, points int) error
+	UpdatePoints(userID uint, points float64) error
+	UpdatePointsWithTransaction(tx interface{}, userID uint, points float64) error
 }
